requirements.txt
torch~=2.0.1+cu117
torchvision~=0.15.2+cu117
matplotlib~=3.7.1
pillow~=10.0.0
tqdm~=4.65.0
numpy~=1.24.3